:output_interval_timeout option implemented.

+++ timeoutcom.rb
@@ -3,7 +3,7 @@ require 'timeout'
 module TimeoutCommand
   module_function
 
-  def parse_timeout(arg)
+  def parse_timespan(arg)
     case arg
     when Integer, Float
       timeout = arg
@@ -18,7 +18,7 @@ module TimeoutCommand
     when /\A\d+(\.\d+)?(?:d|day)\z/
       timeout = $&.to_f * 60 * 60 * 24
     else
-      raise ArgumentError, "invalid time: #{arg.inspect}"
+      raise ArgumentError, "invalid time span: #{arg.inspect}"
     end
     timeout
   end
@@ -58,9 +58,30 @@ module TimeoutCommand
     process_alive?(-pgid)
   end
 
-  def timeout_command(secs, msgout=STDERR)
-    secs = parse_timeout(secs)
-    if secs < 0
+  def last_output_time
+    last_output_time = [STDOUT, STDERR].map {|f|
+      s = f.stat
+      if s.file?
+        s.mtime
+      else
+        nil
+      end
+    }.compact
+    if last_output_time.empty?
+      nil
+    else
+      last_output_time.max
+    end
+  end
+
+  def timeout_command(command_timeout, opts={})
+    msgout = STDERR
+    command_timeout = parse_timespan(command_timeout)
+    output_interval_timeout = nil
+    if opts[:output_interval_timeout]
+      output_interval_timeout = parse_timespan(opts[:output_interval_timeout])
+    end
+    if command_timeout < 0
       raise TimeoutError, 'no time to run a command'
     end
     pid = fork {
@@ -72,29 +93,54 @@ module TimeoutCommand
     rescue Errno::EACCES # already execed.
     rescue Errno::ESRCH # already exited. (setpgid for a zombie fails on OpenBSD)
     end
+    wait_thread = Thread.new {
+      Process.wait2(pid)[1]
+    }
     begin
-      begin
-        timeout(secs) { Process.wait pid }
-      rescue TimeoutError
-        msgout.puts "timeout: #{secs} seconds exceeds." if msgout
-        Thread.new { Process.wait pid }
+      start_time = Time.now
+      limit_time = start_time + command_timeout
+      command_status = nil
+      while true
+        join_timeout = limit_time - Time.now
+        if join_timeout < 0
+          timeout_reason = "command execution time exceeds #{command_timeout} seconds."
+          break 
+        end
+        if output_interval_timeout and
+           t = last_output_time and
+           (tmp_join_timeout = t + output_interval_timeout - Time.now) < join_timeout
+          join_timeout = tmp_join_timeout
+          if join_timeout < 0
+            timeout_reason = "output interval exceeds #{output_interval_timeout} seconds."
+            break
+          end
+        end
+        if wait_thread.join(join_timeout)
+          command_status = wait_thread.value
+          break
+        end
+      end
+      if command_status
+        return command_status
+      else
+        msgout.puts "timeout: #{timeout_reason}." if msgout
         begin
           Process.kill(0, -pid)
           msgout.puts "timeout: the process group is alive." if msgout
           kill_processgroup(pid, msgout)
         rescue Errno::ESRCH # no process
         end
-        raise
-      rescue Interrupt
-        Process.kill("INT", -pid)
-        raise
-      rescue SignalException
-        Process.kill($!.message, -pid)
-        raise
+        raise TimeoutError
       end
+    rescue Interrupt
+      Process.kill("INT", -pid)
+      raise
+    rescue SignalException
+      Process.kill($!.message, -pid)
+      raise
     ensure
       if processgroup_alive?(pid)
-        msgout.puts "process finished but some descendants remain." if msgout
+        msgout.puts "some descendant processes remain." if msgout
         kill_processgroup(pid, msgout)
       end
     end
